requirement.txt
uvicorn==0.29.0  
FastAPI==0.110.2  
pydantic==2.7.0  
numpy==1.24.3
pylev==1.4.0
python-dotenv==1.0.0  
fast-autocomplete==0.9.0   
pandas==2.2.2  
matplotlib==3.7.2  
seaborn==0.13.0  
scipy==1.11.1  
scikit-learn==1.3.0  


#local packages
-e .
